test general was outdated by test_links_general
